+++ rest/src/main/resources/schema.sql
@@ -4,25 +4,25 @@ DROP TABLE IF EXISTS SHOP;
 DROP TABLE IF EXISTS PRODUCT;
 
 CREATE TABLE IF NOT EXISTS shop (
-    id BIGINT PRIMARY KEY,
+    id INT PRIMARY KEY,
     name VARCHAR(255) UNIQUE
 );
 
 CREATE TABLE IF NOT EXISTS transaction (
-    id BIGINT PRIMARY KEY,
+    id INT PRIMARY KEY,
     date DATE,
-    shop_id BIGINT,
+    shop_id INT,
     FOREIGN KEY (shop_id) REFERENCES shop(id)
 );
 
 CREATE TABLE IF NOT EXISTS product (
-    id BIGINT PRIMARY KEY,
+    id INT PRIMARY KEY,
     name VARCHAR(255) UNIQUE
 );
 
 CREATE TABLE IF NOT EXISTS transaction_details (
-    transaction_id BIGINT,
-    product_id BIGINT,
+    transaction_id INT,
+    product_id INT,
     quantity INT,
     price NUMERIC(38, 2),
     PRIMARY KEY (transaction_id, product_id),
